requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.85.0
h11==0.13.0
httptools==0.5.0
idna==3.4
mysql-connector==2.2.9
mysql-connector-python==8.0.30
protobuf==3.20.1
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.20.4
typing-extensions==4.3.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
